requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
paho-mqtt>=1.6.1
pytest>=6.2.5
requests>=2.26.0